pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["Douglas Saturnino <douglasaturnino@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pre-commit = "^3.7.1"
requests = "^2.32.3"
pytest = "^8.2.2"
requests-mock = "^1.12.1"
beautifulsoup4 = "^4.12.3"
ruff = "^0.4.8"
mysql-connector-python = "^8.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
